rename "models" to "model"
