+++ main-service/src/main/java/ru/practicum/explorewithme/main/mapper/CategoryMapper.java
@@ -2,7 +2,7 @@ package ru.practicum.explorewithme.main.mapper;
 
 import org.mapstruct.Mapper;
 import ru.practicum.explorewithme.main.dto.CategoryDto;
-import ru.practicum.explorewithme.main.models.Category;
+import ru.practicum.explorewithme.main.model.Category;
 
 @Mapper(componentModel = "spring")
 public interface CategoryMapper {
