+++ main-service/src/main/java/ru/practicum/explorewithme/main/model/Category.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewithme.main.models;
+package ru.practicum.explorewithme.main.model;
 
 import jakarta.persistence.*;
 import lombok.*;
