+++ main-service/src/main/java/ru/practicum/explorewithme/main/repository/EventRepository.java
@@ -2,7 +2,7 @@ package ru.practicum.explorewithme.main.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import ru.practicum.explorewithme.main.models.Event;
+import ru.practicum.explorewithme.main.model.Event;
 
 public interface EventRepository extends JpaRepository<Event, Long>, QuerydslPredicateExecutor<Event> {
